+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.0] - 02/19/2020
 
-* TODO: Describe initial release.
+* ViewModel with init and dispose functionality
+* ViewModel lifecycle is managed by ViewModelProvider
+* Extension for BuildContext to find ViewModel
+
+## [0.1.1] - 02/20/2020
+
+* Add tests